update span module
